Use an ubuntu mirror closer to the Jenkins datacenter
rjones@lf pointed out that the current download URL was
exceptionally slow and advised to use this mirror
instead.

Note that we also bumped to a later version of 14.04 due
to availability.  We had explicitly chosen 14.04.1 in the
past for reasons that are no longer applicable, so this
should be fine.

Change-Id: I26c8f76266e95a194802bd2e9be5d002206488eb
Signed-off-by: Gregory Haskins <gregory.haskins@gmail.com>